pyproject.toml
[tool.poetry]
name = "wildberries-analytics"
version = "0.1.0"
description = ""
authors = ["a12345 <shu.m2010@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
pandas = "^2.2.3"
matplotlib = "^3.9.2"
seaborn = "^0.13.2"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
